python_api/requirements.txt
flask==1.1.2
flask-sqlalchemy==2.4.4
psycopg2==2.8.5
waitress==1.4.4
flask_login==0.5.0
flask-restx==0.2.0
Flask-Cors==3.0.8
PyJWT==1.7.1